requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.6.2            # via slackclient
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
chardet==3.0.4            # via aiohttp
click==7.1.1              # via pip-tools
django==2.2.11            # via -r requirements.in
idna==2.9                 # via yarl
multidict==4.7.5          # via aiohttp, yarl
pip-tools==4.5.1          # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
pytz==2019.3              # via django
six==1.14.0               # via pip-tools
slackclient==2.5.0        # via -r requirements.in
sqlparse==0.3.1           # via django
yarl==1.4.2               # via aiohttp
